+++ libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: hamaarou <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/30 15:39:31 by hamaarou          #+#    #+#             */
-/*   Updated: 2022/11/01 08:50:24 by hamaarou         ###   ########.fr       */
+/*   Updated: 2022/11/01 18:52:18 by hamaarou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 int		ft_putchar(char c);
 int		ft_putstr(char *str);
 int		count_nbr(long nbr, int bs);
-int		ft_putnbr(int nbr);
+int		ft_putnbr(unsigned long nbr);
 int		ft_putnbr_hex(unsigned long nbr, char bs);
 
 size_t	ft_strlen(char *str);
